+++ test_vqueue/src/main.rs
@@ -35,9 +35,9 @@ fn main() {
             let handle = std::thread::spawn(move || {
                 for i in 0..DATA_PER_WORKER {
                     let data = IPCItem {
-                        sender: worker_id,
+                        sender: worker_id as u64,
                         msg_type: 0,
-                        data: [i; 8],
+                        data: [i as u64; 8],
                     };
                     push(queue_id, data).expect(
                         format!(
